_TO: double to parameter
git-svn-id: http://localhost/svn/entropybroker/trunk@139 6e73ccbf-f98f-4734-8601-04fa8acba097

+++ utils.cpp
@@ -67,9 +67,9 @@ int READ(int fd, char *whereto, size_t len)
 	return cnt;
 }
 
-int READ_TO(int fd, char *whereto, size_t len, int to)
+int READ_TO(int fd, char *whereto, size_t len, double to)
 {
-	double end_ts = get_ts() + (double)to;
+	double end_ts = get_ts() + to;
 	ssize_t cnt=0;
 
 	while(len>0)
@@ -157,9 +157,9 @@ int WRITE(int fd, char *whereto, size_t len)
 	return cnt;
 }
 
-int WRITE_TO(int fd, char *whereto, size_t len, int to)
+int WRITE_TO(int fd, char *whereto, size_t len, double to)
 {
-	double end_ts = get_ts() + (double)to;
+	double end_ts = get_ts() + to;
 	ssize_t cnt=0;
 
 	while(len>0)
