+++ utils.h
@@ -4,9 +4,9 @@
 double get_ts(void);
 long double get_ts_ns(void);
 int READ(int fd, char *whereto, size_t len);
-int READ_TO(int fd, char *whereto, size_t len, int to);
+int READ_TO(int fd, char *whereto, size_t len, double to);
 int WRITE(int fd, char *whereto, size_t len);
-int WRITE_TO(int fd, char *whereto, size_t len, int to);
+int WRITE_TO(int fd, char *whereto, size_t len, double to);
 int start_listen(char *adapter, int portnr, int listen_queue_size);
 int connect_to(char *host, int portnr);
 void disable_nagle(int fd);
